+++ RELEASE_NOTES.md
@@ -17,6 +17,7 @@ v1.0.0-alpha.X
   [#340](https://github.com/OpenAssetIO/OpenAssetIO/issues/340)
   - `SpecificationBase.kTraitIds` to `kTraitSet`.
   - `TraitsData::TraitIds` to `TraitSet`
+  - `TraitsData::traitIds()` to `traitSet()`
 
 
 ### Improvements